Add files via upload
This code takes existing matlab *.mat files for the Trinidad head glider line and creates a new matfile with a structure called bindata that is in the format required for Dan Rudnick's code.   Note the main thing the code does is compute depth averages for the glider data so that the data is on a 10 m grid in the vertical.