Add files via upload
This function creates a series of map structures.  It loads map_XX_AnomOnly.mat and anncycXX.mat where XX is a reference to the glider line.  XX=Tr for the Trinidad line but for other lines might be 66 or 80 etc.  This was originally written by K. Zaba. from SCRIPPS.  I save the output as mapXX.mat  for use with the plotting functions.